Fix warnings on Windows

+++ trantor/utils/MsgBuffer.cc
@@ -150,7 +150,7 @@ ssize_t MsgBuffer::readFd(int fd, int *retErrno)
     struct iovec vec[2];
     size_t writable = writableBytes();
     vec[0].iov_base = begin() + tail_;
-    vec[0].iov_len = writable;
+    vec[0].iov_len = static_cast<int>(writable);
     vec[1].iov_base = extBuffer;
     vec[1].iov_len = sizeof(extBuffer);
     const int iovcnt = (writable < sizeof extBuffer) ? 2 : 1;
